refactor($step1): split test classes between entity and repository.
