+++ src/test/java/qna/domain/QuestionRepositoryTest.java
@@ -1,0 +1,64 @@
+package qna.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import java.util.List;
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@DataJpaTest
+public class QuestionRepositoryTest {
+    @Autowired
+    QuestionRepository questionRepository;
+
+    @Test
+    @DisplayName("save 테스트")
+    void saveTest() {
+        // 데이터 테스트
+        assertThat(QuestionTest.QUESTION1.getId()).isNull();
+        Question actualQuestion1 = questionRepository.save(QuestionTest.QUESTION1);
+        assertThat(actualQuestion1.getId()).isNotNull();
+        assertThat(actualQuestion1.getTitle()).isEqualTo(QuestionTest.QUESTION1.getTitle());
+        assertThat(actualQuestion1.getContents()).isEqualTo(QuestionTest.QUESTION1.getContents());
+        assertThat(actualQuestion1.getCreatedAt()).isNotNull();
+        assertThat(actualQuestion1.getUpdatedAt()).isNotNull();
+
+        // 리스트 테스트
+        Question actualQuestion2 = questionRepository.save(QuestionTest.QUESTION2);
+        List<Question> questionList = questionRepository.findAll();
+        assertThat(questionList.size()).isEqualTo(2);
+        assertThat(questionList).containsExactly(actualQuestion1, actualQuestion2);
+    }
+
+    @Test
+    @DisplayName("전체 검색 테스트")
+    void findByDeletedFalseTest() {
+        questionRepository.save(QuestionTest.QUESTION1);
+        questionRepository.save(QuestionTest.QUESTION2);
+        List<Question> expectedList = questionRepository.findAll();
+        List<Question> actualList = questionRepository.findByDeletedFalse();
+
+        //findByDeletedFalse test
+        assertThat(actualList.size()).isEqualTo(2);
+        assertThat(actualList).isEqualTo(expectedList);
+
+        //deleted false test
+        for(Question question : actualList) {
+            assertThat(question.isDeleted()).isEqualTo(false);
+        }
+    }
+
+    @Test
+    @DisplayName("id 기준 검색 테스트")
+    void findByIdAndDeletedFalse() {
+        Question expected = questionRepository.save(QuestionTest.QUESTION1);
+        Optional<Question> actual = questionRepository.findByIdAndDeletedFalse(expected.getId());
+
+        assertThat(actual.get()).isEqualTo(expected);
+        assertThat(actual.get().isDeleted()).isFalse();
+    }
+}
